workflows: Generalize GitHub action for IPsec key rotations
The GitHub action for IPsec key rotations currently assumes the cluster
has two nodes and that both IPv4 and IPv6 are enabled. These assumptions
don't hold for the EKS conformance test where IPv6 is disabled and three
nodes are created.

This commit therefore adds two new arguments to the GitHub action, to
support IPv6 being enabled/disabled and a variable number of nodes.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>